running Z3 via child processes; TODO suffering from weird JSON bug
