+++ website/package.json
@@ -29,6 +29,7 @@
     "z3-solver": "^4.8.17"
   },
   "devDependencies": {
+    "@docusaurus/eslint-plugin": "^2.0.0-beta.21",
     "@docusaurus/module-type-aliases": "^2.0.0-beta.21",
     "@tsconfig/docusaurus": "^1.0.6",
     "typescript": "^4.7.4"
